Recreate parks dataset
Instead of using the geojson Nico downloaded and manipulated manually, we are
working directly from NYC's open data API. I also made a change by adding
cemeteries and combining shapes that belong to a common park, and removed
tennis courts and baseball diamonds, etc. that are part of a larger facility.