Revise shortest path calculator
Apparently we need to do a few different things to make the osmnx package do what we need it to do. First, the euclidean calculation needs to happen with projected coordinates, and the route needs to happen on a projected network. BUT.... the nearest node lookup has to happen on the lat/long network with lat/long coordinates. Very confusing. 

I did take this opportunity, however, to successfully implement multicore processing in the script. You need about 1.5 GB per thread to store the network information, but Beast shouldn't have any issues running that.